Use realpath/dirname to get real path to shell dir
Magento composer installer uses a symlink strategy for linking shell commands
to their target directory. Previously it was assumed that
basedir/shell/abstract.php was in the current path of the shell script
which is not true if the resque.php shell script is not put in /shell or
symlinked.

This fix assumes standard composer vendor directory structure and recurses
upwards to find the shell dir.